Make tests work on PHP 8
In older versions of PHPUnit there's a "Match" interface, and "match" is a reserved keyword in PHP 8.
Because of this the tests wouldn't work. This commit adds a  really terrible hack around that,
renamning the Match interface in PHPUnit and remapping the files via Composer.

The goal for this is to be temporary until we're on a newer version of PHPUnit
and allow the 3.x branch to be run and tested on PHP 8.